requirements.txt
fastapi[standard]>=0.112.2
starlette>=0.40.0
jinja2>=3.1.4
setuptools>=74.0.0
sqlalchemy>=2.0.32
aiosqlite>=0.20.0
python-dotenv>=1.0.1
pydantic~=2.9.0b2
pyjwt~=2.9.0
passlib[bcrypt]~=1.7.4
bcrypt==4.0.1
httpx~=0.27.2
python-jose~=3.3.0
asgi-ratelimit[full]~=0.10.0
itsdangerous~=2.2.0
pytest~=8.3.3